Add interactive word-wrapping demo
The goal of the demo is to let you to play with the word-wrapping
options in a very simple text editor. You can change the wrapping
width, you can toggle `break_words` and you can change the
`WordSplitter` used.

Clone the `textwrap` repository and use

    cargo run --example interactive

to run the demo. It only work on Unix systems right now, but please
let me know if you want to help port it to Windows.