Back-port interactive demo and add a const fn splitter method.
This commit back-ports the interactive demo from #218 to v0.12.
Additionally, it adds `splitter` method, that clones `&self` and replaces
the splitter with the given one. This made it conforming to become a
`const fn`, allowing `Wrapper::new(w).splitter(s)` idiom in const & static
context, as well as allowing to port the interactive demo.